requirements.txt
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.3.0
rumps>=0.4.0
jinja2>=3.1.2
pytz>=2023.3
# Note: rumps requires pyobjc on macOS. It is usually installed automatically as a dependency.
# If you encounter issues, uncomment the following line:
# pyobjc-framework-Cocoa>=9.0
